Refactoring/cleanup
