+++ src/doc/en/reference/cryptography/index.rst
@@ -34,7 +34,7 @@ Cryptography
    sage/crypto/lattice
    sage/crypto/lwe
 
-   sage/crypto/key_exchange/key_exchange
+   sage/crypto/key_exchange/key_exchange_scheme
    sage/crypto/key_exchange/diffie_hellman
 
 .. include:: ../footer.txt
