+++ src/sage/crypto/all.py
@@ -20,4 +20,9 @@ lazy_import('sage.crypto.lfsr', ['lfsr_sequence',
                                  'lfsr_autocorrelation',
                                  'lfsr_connection_polynomial',
                                  ])
+
+lazy_import('sage.crypto.key_exchange.all', ['KeyExchangeScheme',
+                                 'DiffieHellman',
+                                 ])
+
 del lazy_import
