+++ src/sage/crypto/key_exchange/all.py
@@ -1,4 +1,2 @@
-from sage.misc.lazy_import import lazy_import
-
-lazy_import('sage.crypto.key_exchange.diffie_hellman', 'DiffieHellman')
-del lazy_import
+from sage.crypto.key_exchange.diffie_hellman import DiffieHellman
+from sage.crypto.key_exchange.key_exchange_scheme import KeyExchangeScheme
